Add GUI window with evaluation result and refresh option
-Replace program output from console to window using WinForms.

-Add rate-limited concurrent API calls to avoid 429 errors

+++ InventoryEvaluator/Models/EvaluatedItem.cs
@@ -4,9 +4,9 @@ public class EvaluatedItem
 {
     public string Name { get; set; } = "";
     public int Amount { get; set; }  
-    public double AvgSellPrice { get; set; } 
-    public double HighestBuyPrice { get; set; }
-    public double TotalSellPrice { get; set; }
+    public int AvgSellPrice { get; set; } 
+    public int HighestBuyPrice { get; set; }
+    public int TotalSellPrice { get; set; }
 
     public override string ToString()
     {
